Re-structuring code architecture
