+++ api/handlers/home.go
@@ -1,0 +1,7 @@
+package handlers
+
+import "github.com/gofiber/fiber/v2"
+
+func Home(c *fiber.Ctx) error {
+	return c.SendString("Hello, Gym-People-ESGI!")
+}
