--- fixtures/sessions.yml
@@ -1,17 +1,0 @@
-- id: 1
-  name: "Morning Session"
-  description: "Start your day with a burst of energy"
-  group_id: 1
-  type: "Yoga"
-
-- id: 2
-  name: "Afternoon Session"
-  description: "Keep your momentum going"
-  group_id: 2
-  type: "Muscu"
-
-- id: 3
-  name: "Evening Session"
-  description: "Wind down and relax"
-  group_id: 3
-  type: "Cardio"
