--- handlers/login.go
@@ -1,78 +1,0 @@
-package handlers
-
-import (
-	"os"
-	"time"
-
-	"github.com/Honzikoi/gym-challenge/database"
-	"github.com/Honzikoi/gym-challenge/models"
-	"github.com/gofiber/fiber/v2"
-	"github.com/golang-jwt/jwt/v4"
-	"golang.org/x/crypto/bcrypt"
-)
-
-// Login godoc
-//	@Summary		Log in a user
-//	@Description	Authenticate a user with email and password
-//	@Tags			Signup & Login
-//	@Accept			json
-//	@Produce		json
-//	@Param			login	body		LoginRequest	true	"Login request"
-//	@Success		200		{string}	JWT				token
-//	@Failure		400		{string}	Invalid			request	payload
-//	@Failure		401		{string}	Invalid			email	or			password
-//	@Failure		500		{string}	Failed			to		generate	token
-//	@Router			/login [post]
-type LoginRequest struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-func Login(c *fiber.Ctx) error {
-	db := database.DB.Db
-
-	var req LoginRequest
-	if err := c.BodyParser(&req); err != nil {
-		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
-			"error": "Invalid request payload",
-		})
-	}
-
-	var user models.Users
-	if err := db.Where("email = ?", req.Email).First(&user).Error; err != nil {
-		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
-			"error": "Invalid email or password",
-		})
-	}
-
-	if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(req.Password)); err != nil {
-		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
-			"error": "Invalid email or password",
-		})
-	}
-
-	token, err := generateJWT(user)
-	if err != nil {
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"error": "Failed to generate token",
-		})
-	}
-
-	return c.JSON(fiber.Map{
-		"token": token,
-	})
-}
-
-func generateJWT(user models.Users) (string, error) {
-	jwtSecret := []byte(os.Getenv("JWT_SECRET"))
-
-	claims := jwt.MapClaims{
-		"authorized": true,
-		"role":       user.Role.Name,
-		"user_id":    user.ID,
-		"exp":        time.Now().Add(time.Hour * 24).Unix(),
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString(jwtSecret)
-}
