--- handlers/profile.go
@@ -1,15 +1,0 @@
-package handlers
-
-import (
-	"fmt"
-
-	"github.com/gofiber/fiber/v2"
-	"github.com/golang-jwt/jwt/v4"
-)
-
-func Profile(c *fiber.Ctx) error {
-	user := c.Locals("user").(*jwt.Token)
-	claims := user.Claims.(jwt.MapClaims)
-	email := claims["email"].(string)
-	return c.SendString(fmt.Sprintf("Welcome to the protected area, %s!", email))
-}
